Implement required code changes for common target to compile as C++
* Add casts where necessary mostly for void * -> any other type
* Replaced use of min, max, swap macros with std variant
* Fixed where `const char` strings were passed/stored to non-const variables
* Updated VECTOR_RESIZE with C++ variant that relays on decltype to make type safe cast from memory manager

+++ src/common/db.c
@@ -2430,7 +2430,7 @@ static enum DBOptions db_fix_options(enum DBType type, enum DBOptions options)
 			return (enum DBOptions)(options&~(DB_OPT_DUP_KEY|DB_OPT_RELEASE_KEY));
 
 		default:
-			ShowError("db_fix_options: Unknown database type %u with options %x\n", type, options);
+			ShowError("db_fix_options: Unknown database type %u with options %x\n", (unsigned int)type, (unsigned int)options);
 			FALLTHROUGH
 		case DB_STRING:
 		case DB_ISTRING: // String databases, no fix required
@@ -2461,7 +2461,7 @@ static DBComparator db_default_cmp(enum DBType type)
 		case DB_INT64:   return &db_int64_cmp;
 		case DB_UINT64:  return &db_uint64_cmp;
 		default:
-			ShowError("db_default_cmp: Unknown database type %u\n", type);
+			ShowError("db_default_cmp: Unknown database type %u\n", (unsigned int)type);
 			return NULL;
 	}
 }
@@ -2489,7 +2489,7 @@ static DBHasher db_default_hash(enum DBType type)
 		case DB_INT64:   return &db_int64_hash;
 		case DB_UINT64:  return &db_uint64_hash;
 		default:
-			ShowError("db_default_hash: Unknown database type %u\n", type);
+			ShowError("db_default_hash: Unknown database type %u\n", (unsigned int)type);
 			return NULL;
 	}
 }
@@ -2545,7 +2545,7 @@ static DBReleaser db_custom_release(enum DBReleaseOption which)
 		case DB_RELEASE_DATA:    return &db_release_data;
 		case DB_RELEASE_BOTH:    return &db_release_both;
 		default:
-			ShowError("db_custom_release: Unknown release options %u\n", which);
+			ShowError("db_custom_release: Unknown release options %u\n", (unsigned int)which);
 			return NULL;
 	}
 }
